+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ## [0.10.2] - 2025-03-01
+### Fixed
+- Repeated marker renderings on page load
 
 ## [0.10.1] - 2025-03-01