+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+- Disable clustering at zoom level 12
+
 ## [0.11.0] - 2025-03-01
 
 ### Added